fix(gptdriver): address PR review comments from gdeluna-branch
- Revert CI workflow actions/checkout@v3 back to v5 (unintended downgrade)
- Extract mobileboost_ids to centralized mobileboost-ids.json config
- Revert AutoDeepLinkTestActivity if(false) debugging artifact
- Optimize LogOutputActivity with StringBuilder.insert(0,...) for reversed logs
- Extract test utilities from MainActivity to TestBedHelper class
- Replace deprecated InstrumentationRegistry with platform.app variant
