+++ .github/workflows/unit-and-instrumented-tests-action.yml
@@ -13,7 +13,7 @@ jobs:
     runs-on: macos-latest
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v5
       # repo's gradle is configured to run on java 17
       - name: Setup java 17 for gradle
         uses: actions/setup-java@v5
@@ -36,7 +36,7 @@ jobs:
     runs-on: macos-latest
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v5
       # create an emulator with google apis, runs on java 8
       - name: Create Android emulator
         run: |
@@ -76,7 +76,7 @@ jobs:
             runs-on: macos-latest
             steps:
               - name: Checkout
-                uses: actions/checkout@v3
+                uses: actions/checkout@v5
               # create an emulator with google apis, runs on java 8
               - name: Create Android emulator
                 run: |
@@ -112,7 +112,7 @@ jobs:
     runs-on: macos-latest
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v5
       # repo's gradle is configured to run on java 17
       - name: Setup java 17 for gradle
         uses: actions/setup-java@v5
@@ -134,7 +134,7 @@ jobs:
     runs-on: macos-latest
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v5
       # create an emulator with google apis, runs on java 8
       - name: Create Android emulator
         run: |
