+++ Branch-SDK-GPTDriver/mobileboost/dependencies/enable-app-link.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android Dependency] Enable App Link",
-  "mobileboost_id": "gMTZpITvB9ejIJTtbZPd",
   "tags": [],
   "steps": [
     {
