+++ Branch-SDK-GPTDriver/mobileboost/dependencies/force-close-app-and-enable-app-link.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android Dependency] Force Close App and Enable App Link",
-  "mobileboost_id": "bbZzyJWXDMT9WQUVpFMT",
   "tags": [],
   "steps": [
     {
