+++ Branch-SDK-GPTDriver/mobileboost/dependencies/force-close-app.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android Dependency] Force Close App",
-  "mobileboost_id": "b86SEJYY8io8sNbkbAnR",
   "tags": [],
   "steps": [
     {
