+++ Branch-SDK-GPTDriver/mobileboost/suites/async-identity-management.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android] Async Identity Management",
-  "mobileboost_id": "1esSo4Nd29pYdrEykd5j",
   "tags": [
     "Android",
     "SDK Rewrite"
