+++ Branch-SDK-GPTDriver/mobileboost/suites/event-tracking-post-sdk-rewrite.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android] Event Tracking - Post SDK Rewrite",
-  "mobileboost_id": "37oD8BqkAEInKrXnUTPr",
   "tags": [
     "Android",
     "SDK Rewrite"
