+++ Branch-SDK-GPTDriver/mobileboost/suites/log-event-custom.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android] Log Event - Custom",
-  "mobileboost_id": "VhY0ZlKiIIHPlOs9Ql87",
   "tags": [
     "Android",
     "Release"
