+++ Branch-SDK-GPTDriver/mobileboost/suites/log-event-standard.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android] Log event - Standard",
-  "mobileboost_id": "C34I7gpuPhg4Z1mLAFD3",
   "tags": [
     "Android",
     "Release"
