+++ Branch-SDK-GPTDriver/mobileboost/suites/performance-verify-ad-fields.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android] Performance - Verify Ad Fields",
-  "mobileboost_id": "kAdwWjzdT5TV93m24mFi",
   "tags": [],
   "steps": [
     {
