+++ Branch-SDK-GPTDriver/mobileboost/suites/set-customer-event-alias.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android] Set Customer Event Alias",
-  "mobileboost_id": "Fv7Yl2Fb6zVuNypF0Q3Q",
   "tags": [
     "Android"
   ],
