+++ Branch-SDK-GPTDriver/mobileboost/suites/set-session-id.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android] Set Session ID",
-  "mobileboost_id": "jEHKVaiT8ci9cYRM0rdl",
   "tags": [
     "Android"
   ],
