+++ Branch-SDK-GPTDriver/mobileboost/suites/uri-scheme-cold-open.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android] URI Scheme - Cold Open",
-  "mobileboost_id": "vyTzci1niNmvAf8vRxp7",
   "tags": [
     "Android",
     "Flaky"
