+++ Branch-SDK-GPTDriver/mobileboost/suites/uri-scheme-warm-open.json
@@ -1,6 +1,5 @@
 {
   "name": "[Android] URI Scheme - Warm Open",
-  "mobileboost_id": "tjzKXu9T6pTMPoOUkC64",
   "tags": [
     "Android",
     "Flaky"
