+++ Branch-SDK-GPTDriver/src/main/java/io/branch/gptdriver/BaseGptDriverTest.kt
@@ -34,8 +34,7 @@ abstract class BaseGptDriverTest {
     open fun setUp() {
         val apiKey = BuildConfig.MOBILEBOOST_API_KEY.let { key ->
             key.ifEmpty {
-                @Suppress("DEPRECATION")
-                androidx.test.InstrumentationRegistry.getArguments()
+                androidx.test.platform.app.InstrumentationRegistry.getArguments()
                     .getString("GPTDRIVER_API_KEY") ?: ""
             }
         }
