+++ Branch-SDK-GPTDriver/src/main/java/io/branch/gptdriver/tests/DeepLinkWarmOpenHybridTest.kt
@@ -51,8 +51,7 @@ class DeepLinkWarmOpenHybridTest {
     fun setUp() {
         val apiKey = BuildConfig.MOBILEBOOST_API_KEY.let { key ->
             key.ifEmpty {
-                @Suppress("DEPRECATION")
-                androidx.test.InstrumentationRegistry.getArguments()
+                androidx.test.platform.app.InstrumentationRegistry.getArguments()
                     .getString("GPTDRIVER_API_KEY") ?: ""
             }
         }
